+++ Sources/PSQLKit/TestSQLDatabase.swift
@@ -1,0 +1,22 @@
+import SQLKit
+import PostgresKit
+
+final class TestSQLDatabase: SQLDatabase {
+    let logger: Logger
+    let eventLoop: EventLoop
+    var results: [String]
+    let dialect: SQLDialect = PostgresDialect()
+    
+    init() {
+        self.logger = .init(label: "com.hiimtmac.psqlkit")
+        self.eventLoop = EmbeddedEventLoop()
+        self.results = []
+    }
+    
+    func execute(sql query: SQLExpression, _ onRow: @escaping (SQLRow) -> ()) -> EventLoopFuture<Void> {
+        var serializer = SQLSerializer(database: self)
+        query.serialize(to: &serializer)
+        results.append(serializer.sql)
+        return self.eventLoop.makeSucceededFuture(())
+    }
+}
